FracCalc Ok

+++ pyphs/tests/test_RLC_python.py
@@ -4,21 +4,26 @@ Created on Sat May 21 10:50:30 2016
 
 @author: Falaize
 """
+from __future__ import absolute_import, division, print_function
+import pyphs
+import os
 
 
 def label():
     """
     System's netlist and folder label
     """
-    return "RLC"
+    return "rlc"
 
 
-def netlist_filename(phs=None):
-    import os
-    if phs is None:
-        return os.getcwd() + os.sep + label() + '.net'
-    else:
-        return phs.path + os.sep + label() + '.net'
+def path():
+    """
+    System's netlist and folder path
+    """
+    p = os.getcwd() + os.sep + label()
+    if not os.path.exists(p):
+        os.makedirs(p)
+    return p
 
 
 def samplerate():
@@ -28,20 +33,21 @@ def samplerate():
     return 96e3
 
 
-def write_netlist(phs, R1=1e3, L=5e-2, C=2e-6):
+def netlist(R1=1e3, L=5e-2, C=2e-6):
     """
     Write netlist for RLC circuit
     """
+    netlist = pyphs.PHSNetlist(path() + os.sep + label() + '.net', clear=True)
 
-    datum = phs.graph.netlist.datum
+    datum = netlist.datum
 
     # input voltage
     source = {'dictionary': 'electronics',
               'component': 'source',
               'label': 'IN',
               'nodes': (datum, 'A'),
-              'arguments': {'type': "'voltage'"}}
-    phs.graph.netlist.add_line(source)
+              'arguments': {'type': 'voltage'}}
+    netlist.add_line(source)
 
     # resistor 1
     resistance = {'dictionary': 'electronics',
@@ -49,7 +55,7 @@ def write_netlist(phs, R1=1e3, L=5e-2, C=2e-6):
                   'label': 'R1',
                   'nodes': ('A', 'B'),
                   'arguments': {'R': ('R1', R1)}}
-    phs.graph.netlist.add_line(resistance)
+    netlist.add_line(resistance)
 
     # inductor
     inductor = {'dictionary': 'electronics',
@@ -57,7 +63,7 @@ def write_netlist(phs, R1=1e3, L=5e-2, C=2e-6):
                 'label': 'L',
                 'nodes': ('B', 'C'),
                 'arguments': {'L': ('L', L)}}
-    phs.graph.netlist.add_line(inductor)
+    netlist.add_line(inductor)
 
     # capacitor
     capacitor = {'dictionary': 'electronics',
@@ -65,26 +71,18 @@ def write_netlist(phs, R1=1e3, L=5e-2, C=2e-6):
                  'label': 'C',
                  'nodes': ('C', datum),
                  'arguments': {'C': ('C', C)}}
-    phs.graph.netlist.add_line(capacitor)
-
-    phs.graph.netlist.write(filename=netlist_filename(phs))
+    netlist.add_line(capacitor)
+    return netlist
 
 
 def init_phs():
     # import pHobj
-    from pyphs import PortHamiltonianObject
-    import os
-    phs = PortHamiltonianObject(label=label(),
-                                path=os.getcwd() + os.sep + label())
+    phs = pyphs.PHSObject(netlist(), label=label(), path=path())
     return phs
 
 
-def build_graph(phs):
-    phs.build_from_netlist(netlist_filename(phs))
-
-
 def input_sequence(amp=100., f0=100.):
-    from pyphs.misc.signals.synthesis import signalgenerator
+    from pyphs import signalgenerator
     fs = samplerate()
     nsin = int(2*fs/f0)
     SigIn = signalgenerator(which="sin", n=nsin, ramp_on=False,
@@ -100,29 +98,20 @@ def input_sequence(amp=100., f0=100.):
 def simulation(phs):
     opts = {'fs': samplerate(),
             'language': 'python',
-            'split': False}
+            'split': True,
+            'progressbar': True}
     u, nt = input_sequence()
-    phs.simu.init(sequ=u, nt=nt, opts=opts)
-    phs.simu.process()
+    phs.Simu.config.update(opts)
+    phs.Simu.init(sequ=u, nt=nt)
+    phs.Simu.process()
 
 
-def gen_cpp(phs):
-    phs.cppbuild()
-    phs.cpp.gen_main()
-    phs.cpp.gen_phobj()
-    phs.cpp.gen_data()
-
-
-def run_test():
+def run_test(clean=True):
     phs = init_phs()
-    write_netlist(phs)
-    build_graph(phs)
     simulation(phs)
-    gen_cpp(phs)
-#    phs.plot_powerbal()
-#    phs.plot_data([('x', 0), ('dx', 1), ('dxH', 1)])
-    import shutil
-    shutil.rmtree(phs.path, ignore_errors=True)
+    if clean:
+        import shutil
+        shutil.rmtree(phs.path, ignore_errors=True)
     return True
 
 if __name__ is '__main__':
